api/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
kafka-python==2.0.2
mysql-connector-python==8.0.26